web-app/src/test/java/com/ram/app/controllers/LoginControllerTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */

package com.ram.app.controllers;

import com.ram.app.model.LoginModel;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;
import org.springframework.validation.BindingResult;
import org.springframework.web.servlet.ModelAndView;

/**
 *
 * @author ramdangi
 */
public class LoginControllerTest {
    
    public LoginControllerTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    /**
     * Test of init method, of class LoginController.
     */
    @Test
    public void testInit() {
        System.out.println("init");
        LoginController instance = new LoginController();
        ModelAndView expResult = null;
        ModelAndView result = instance.init();
        //assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        //fail("The test case is a prototype.");
    }

    /**
     * Test of login method, of class LoginController.
     */
    @Test
    public void testLogin() {
        System.out.println("login");
        LoginModel loginModel = null;
        BindingResult bindingResult = null;
        HttpServletRequest request = null;
        HttpServletResponse response = null;
        LoginController instance = new LoginController();
        ModelAndView expResult = null;
        ModelAndView result = instance.login(loginModel, bindingResult, request, response);
        //assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        //fail("The test case is a prototype.");
    }
    
}
